move files out of resolve dir
