+++ packages/lib/src/offchainResolver/index.ts
@@ -1,7 +1,7 @@
-export { decodeCalldata } from './resolve/decodeCalldata';
-export { encodeUserProfile } from './resolve/encodeUserProfile';
-export { resolveWithProof } from './resolve/resolveWithProof';
+export { decodeCalldata } from './decodeCalldata';
+export { encodeUserProfile } from './encodeUserProfile';
+export { resolveWithProof } from './resolveWithProof';
 export type {
     CiipResponse as ResolveResponse,
     DecodedCcipRequest,
-} from './resolve/types';
+} from './types';
